% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydata.R
\docType{data}
\name{mylist}
\alias{mylist}
\title{Data for R Shiny App}
\format{
A list of lists, each with length 2: a dataframe and a confidence interval
\describe{
  \item{dataframe}{a dataframe of X and Y vales}
  \item{CI}{A confidence interval from appropriate t test}
  ...
}
}
\source{
Custom data
}
\usage{
mylist
}
\description{
A list of lists containing data for use in an R Shiny App
}
\keyword{datasets}
