% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor_Function.R
\name{myttest}
\alias{myttest}
\title{Constructor Function}
\usage{
myttest(x, y, alpha = 0.05, paired = FALSE)
}
\arguments{
\item{x}{The x variable data}

\item{y}{The y variable data}

\item{alpha}{an alpha value, normally set to 0.05}

\item{paired}{Is data paired or unpaired}
}
\value{
A list containing: test results, a dataframe of the two combined datasets, confidence interval and P-Value
}
\description{
Takes in two sets of data and produces a list containing: test results,
a dataframe, and the confidence interval and P - value
}
\examples{
\dontrun{myttest(x, y, alpha = 0.05, paired = FALSE)}
}
