% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Function.R
\name{print.rttest}
\alias{print.rttest}
\title{Print Function}
\usage{
\method{print}{rttest}(x, ...)
}
\arguments{
\item{x}{A variable}

\item{...}{A table variable}
}
\value{
A Confidence Intervale
}
\description{
Prints Results of CI
}
\examples{
\dontrun{print(obj)}
}
