% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_Function.R
\name{shinyfunction}
\alias{shinyfunction}
\title{Dynamic plotting and printing of Boxplot and Confidence Interval}
\usage{
shinyfunction()
}
\value{
Boxplot and Confidence Interval
}
\description{
A website-like app that is used to select one of the datasets and plot a boxplot and confidence interval
}
\examples{
\dontrun{shinyplots()}
}
