% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor_Function.R
\name{myttest}
\alias{myttest}
\title{Constructor Function}
\usage{
myttest(x, y, alpha = 0.05, paired = FALSE)
}
\arguments{
\item{x}{The x dataset}

\item{y}{The y dataset}

\item{alpha}{an alpha value}

\item{paired}{Paired Data}
}
\value{
A list
}
\description{
Takes in Data and produces a list
}
\examples{
\dontrun{myttest(x, y, alpha = 0.05, paired = FALSE)}
}
