% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Function.R
\name{plot.rttest}
\alias{plot.rttest}
\title{Plot Function}
\usage{
\method{plot}{rttest}(x, ...)
}
\arguments{
\item{x}{A variable}

\item{...}{A variable}
}
\value{
A boxplot of data from myttest
}
\description{
S3 method for plotting results of data from myttest
}
\examples{
\dontrun{plot(obj)}
}
